requirements.txt
autopep8>=1.5.7
click>=8.0.1
cycler>=0.10.0
decorator>=4.4.2
Flask>=2.0.1
Flask-PyMongo>=2.3.0
imageio>=2.9.0
itsdangerous>=2.0.1
Jinja2>=3.0.1
joblib>=1.0.1
kiwisolver>=1.3.1
MarkupSafe>=2.0.1
matplotlib>=3.4.2
networkx>=2.5.1
numpy>=1.20.3
opencv-python>=4.5.2.52
pandas>=1.2.4
Pillow>=8.2.0
pycodestyle>=2.7.0
pymongo>=3.11.4
pyparsing>=2.4.7
python-dateutil>=2.8.1
pytz>=2021.1
PyWavelets>=1.1.1
scikit-image>=0.18.1
scikit-learn>=0.24.2
scipy>=1.6.3
six>=1.16.0
threadpoolctl>=2.1.0
tifffile>=2021.4.8
toml>=0.10.2
torch>=1.8.1+cu111
torchaudio>=0.8.1
torchsummary>=1.5.1
torchvision>=0.9.1+cu111
tqdm>=4.61.0
typing-extensions>=3.10.0.0
Werkzeug>=2.0.1
